requirements.txt
tgcrypto
telethon
asyncio
asyncpg
ffmpeg-python
python-telegram-bot
py-tgcalls
pyrogram
youtube-search-python==1.4.9
yt-dlp==2021.11.10.1
git+https://github.com/ytdl-org/youtube-dl@master
youtube-search
python-dotenv
dnspython
gitpython
aiofiles
aiohttp
requests
heroku3
motor
psutil
future
wget